+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+## 1.0.9
+You can (optionally) declare Collection Service configuration details in your module or component providers:
+```ts
+providers: [
+  {
+    provide: 'COLLECTION_SERVICE_OPTIONS',
+    useValue: {
+      allowFetchedDuplicates: environment.production,
+    }
+  },
+]
+```
+Token `COLLECTION_SERVICE_OPTIONS` is just a string to don't break lazy-loading (if you are using it).
+
+Options structure:
+```ts
+interface CollectionServiceOptions {
+  comparatorFields?: string[];
+  throwOnDuplicates?: string;
+  allowFetchedDuplicates?: boolean; // if not set: true
+}
+```
+
 ## 1.0.8
 * Fix itemViewModel.isProcessing; 
 * Make all selectors in itemViewModel protected from streams without pre-emitted value;